+++ contao/languages/en/tl_list_config_element.php
@@ -1,0 +1,5 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_list_config_element'];
+
+$lang['locationFieldSelector'] = ['Select location field', 'Select the field that contains the location information.'];