+++ src/ContaoLocationBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace HeimrichHannot\LocationBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class ContaoLocationBundle extends Bundle
-{
-}