+++ src/ContaoManager/Plugin.php
@@ -6,14 +6,9 @@ use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
-use Contao\ManagerPlugin\Config\ConfigPluginInterface;
-use Contao\ManagerPlugin\Routing\RoutingPluginInterface;
-use HeimrichHannot\LocationBundle\ContaoLocationBundle;
-use Symfony\Component\Config\Loader\LoaderInterface;
-use Symfony\Component\Config\Loader\LoaderResolverInterface;
-use Symfony\Component\HttpKernel\KernelInterface;
+use HeimrichHannot\LocationBundle\HeimrichHannotLocationBundle;
 
-class Plugin implements BundlePluginInterface, RoutingPluginInterface, ConfigPluginInterface
+class Plugin implements BundlePluginInterface
 {
     /**
      * {@inheritdoc}
@@ -25,24 +20,7 @@ class Plugin implements BundlePluginInterface, RoutingPluginInterface, ConfigPlu
         ];
 
         return [
-            BundleConfig::create(ContaoLocationBundle::class)->setLoadAfter($loadAfter)
+            BundleConfig::create(HeimrichHannotLocationBundle::class)->setLoadAfter($loadAfter)
         ];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getRouteCollection(LoaderResolverInterface $resolver, KernelInterface $kernel)
-    {
-//        return $resolver->resolve(__DIR__.'/../Resources/config/routing.yml')->load(__DIR__.'/../Resources/config/routing.yml');
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig)
-    {
-        $loader->load('@ContaoLocationBundle/Resources/config/services.yml');
-        $loader->load('@ContaoLocationBundle/Resources/config/datacontainers.yml');
     }
 }