+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.0.0] - 2021-10-28
+- Added: Config element type
+- Added: license file
+- Changed: raised contao minimum version to 4.9
+- Changed: refactored to new bundle structure
+- Changed: renamed bundle class
+- Changed: added missing dependencies
+- Removed: test setup
+
 ## [1.0.1] - 2020-07-09
 
 - set user/user group fields to blob